(chour) tests: Dedupped _DummyPreferences, moved into tests/conftest.py
